requirements.txt
antares-client==1.2.0
google-cloud-logging
google-cloud-storage==1.43.0
SQLAlchemy==1.4.27
pg8000==1.23.0
psycopg2-binary
numpy
statsmodels
elasticsearch_dsl
Flask >= 1.0
gunicorn==20.1.0
julian==0.14